Fixed minor Release & fixed point compiler warnings on Linux
http://codereview.appspot.com/6118050/



git-svn-id: http://skia.googlecode.com/svn/trunk@3766 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ samplecode/SampleAnimBlur.cpp
@@ -45,7 +45,7 @@ protected:
         };
         SkRandom random;
 
-        for (int i = 0; i < SK_ARRAY_COUNT(gStyles); ++i) {
+        for (size_t i = 0; i < SK_ARRAY_COUNT(gStyles); ++i) {
             SkMaskFilter* mf = SkBlurMaskFilter::Create(blurRadius,
                                        gStyles[i],
                                        SkBlurMaskFilter::kHighQuality_BlurFlag);
